added routes for all the scraper endpoints, including puppeteer to pull snkrdunk listing data
